+++ packages/database/lib/index.d.ts
@@ -380,7 +380,7 @@ export namespace FirebaseDatabaseTypes {
      */
     transaction(
       // eslint-disable-next-line @typescript-eslint/ban-types
-      transactionUpdate: (currentData: object) => object | undefined,
+      transactionUpdate: (currentData: any) => any | undefined,
       onComplete?: (error: Error | null, committed: boolean, finalResult: DataSnapshot) => void,
       applyLocally?: boolean,
     ): Promise<TransactionResult>;